requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-background-tasks==1.2.8
waitress==2.1.2
requests==2.31.0